refactor posts handlers into modules
