+++ src/features/posts/index.js
@@ -1,9 +1,13 @@
-import { initPostHandlers } from './actions.js';
+import { initCommentHandlers } from './comments.js';
+import { initReactionHandlers } from './reactions.js';
+import { initModerationHandlers } from './moderation.js';
 import { initFilterHandlers, applyFilterAndRender } from './filters.js';
 import { initPreviewHandlers } from './preview.js';
 
 export function initPosts() {
-  initPostHandlers();
+  initCommentHandlers();
+  initReactionHandlers();
+  initModerationHandlers();
   initFilterHandlers();
   initPreviewHandlers();
 }
